Move proprocessing of whitelist-uids option to SanitizeCommands class.
